fixed: Uncaught ReferenceError: provider is not defined
