+++ package.json
@@ -1,0 +1,30 @@
+{
+  "name": "angular-logger",
+  "version": "1.0.0",
+  "description": "> Enhanced $log in AngularJS",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/punknroll/angular-logger.git"
+  },
+  "keywords": [
+    "angular",
+    "logger",
+    "angular-logger"
+  ],
+  "author": "",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/pdorgambide/angular-logger"
+  },
+  "homepage": "https://github.com/pdorgambide/angular-logger",
+  "devDependencies": {
+    "gulp": "^3.8.11",
+    "gulp-clean": "^0.3.1",
+    "gulp-concat": "^2.5.2",
+    "gulp-uglify": "^1.1.0"
+  }
+}
